refactor map utils
