+++ lib/widgets/map/trufi_online_map.dart
@@ -7,7 +7,7 @@ import 'package:trufi_core/blocs/gps_location/location_provider_cubit.dart';
 import 'package:trufi_core/blocs/preferences_cubit.dart';
 
 import '../../trufi_configuration.dart';
-import '../../trufi_map_utils.dart';
+import 'utils/trufi_map_utils.dart';
 import 'trufi_map_controller.dart';
 
 typedef LayerOptionsBuilder = List<LayerOptions> Function(BuildContext context);
