+++ test/trufi_map_util_unit_test.dart
@@ -4,7 +4,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:latlong/latlong.dart';
 
 import 'package:trufi_core/entities/plan_entity/plan_entity.dart';
-import 'package:trufi_core/trufi_map_utils.dart';
+import 'package:trufi_core/widgets/map/utils/trufi_map_utils.dart';
 void main() {
   const double _latitude = 18.0;
   const double _longitude = 129.0;
